moves ConnectWalletToast to components root
